+++ README.md
@@ -91,8 +91,8 @@ npm run lint:fix
 ## Структура проекта
 
 - `src/index.ts` — точка входа приложения.
-- `src/bot/TelegramBot.ts` — логика бота и обработка сообщений.
-- `src/application/use-cases/ai/ChatGPTService.ts` — взаимодействие c OpenAI.
+- `src/infrastructure/telegram/TelegramBot.ts` — логика бота и обработка сообщений.
+- `src/infrastructure/external/ChatGPTService.ts` — взаимодействие c OpenAI.
 - `src/application/use-cases/chat/ChatMemory.ts` — хранение истории и её суммаризация.
 - `dist/` — собранные файлы.
 - Сервисы определяются через интерфейсы и Symbol-ключи Inversify; реализации
