+++ src/infrastructure/config/TestEnvService.ts
@@ -1,7 +1,10 @@
 import { injectable } from 'inversify';
 import { ChatModel } from 'openai/resources/shared';
 
-import { Env, EnvService } from '../../interfaces/env/EnvService.interface';
+import {
+  Env,
+  EnvService,
+} from '../../application/interfaces/env/EnvService.interface';
 import { envSchema } from './envSchema';
 
 @injectable()
