+++ src/infrastructure/persistence/sqlite/DbProvider.ts
@@ -5,17 +5,17 @@ import sqlite3 from 'sqlite3';
 import {
   ENV_SERVICE_ID,
   EnvService,
-} from '../../application/interfaces/env/EnvService.interface';
-import type { Logger } from '../../application/interfaces/logging/Logger.interface';
+} from '../../../application/interfaces/env/EnvService.interface';
+import type { Logger } from '../../../application/interfaces/logging/Logger.interface';
 import {
   LOGGER_FACTORY_ID,
   type LoggerFactory,
-} from '../../application/interfaces/logging/LoggerFactory.interface';
+} from '../../../application/interfaces/logging/LoggerFactory.interface';
 import {
   type DbProvider,
   type SqlDatabase,
-} from '../../domain/repositories/DbProvider.interface';
-import { parseDatabaseUrl } from '../../utils/database';
+} from '../../../domain/repositories/DbProvider.interface';
+import { parseDatabaseUrl } from '../../../utils/database';
 
 export type SQLiteDbProvider = DbProvider;
