+++ src/infrastructure/persistence/sqlite/SQLiteChatUserRepository.ts
@@ -1,11 +1,11 @@
 import { inject, injectable } from 'inversify';
 
-import type { ChatUserRepository } from '../../domain/repositories/ChatUserRepository.interface';
+import type { ChatUserRepository } from '../../../domain/repositories/ChatUserRepository.interface';
 import {
   DB_PROVIDER_ID,
   type DbProvider,
   type SqlDatabase,
-} from '../../domain/repositories/DbProvider.interface';
+} from '../../../domain/repositories/DbProvider.interface';
 
 @injectable()
 export class SQLiteChatUserRepository implements ChatUserRepository {
