+++ src/infrastructure/persistence/sqlite/SQLiteMessageRepository.ts
@@ -1,13 +1,13 @@
 import { inject, injectable } from 'inversify';
 
-import type { ChatMessage } from '../../application/interfaces/ai/AIService.interface';
-import type { StoredMessage } from '../../application/interfaces/messages/StoredMessage.interface';
+import type { ChatMessage } from '../../../application/interfaces/ai/AIService.interface';
+import type { StoredMessage } from '../../../application/interfaces/messages/StoredMessage.interface';
 import {
   DB_PROVIDER_ID,
   type DbProvider,
   type SqlDatabase,
-} from '../../domain/repositories/DbProvider.interface';
-import type { MessageRepository } from '../../domain/repositories/MessageRepository.interface';
+} from '../../../domain/repositories/DbProvider.interface';
+import type { MessageRepository } from '../../../domain/repositories/MessageRepository.interface';
 
 @injectable()
 export class SQLiteMessageRepository implements MessageRepository {
