+++ test/ChatGPTService.test.ts
@@ -2,8 +2,8 @@ import { promises as fs } from 'fs';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import type { ChatMessage } from '../src/application/interfaces/ai/AIService.interface';
-import type { ChatGPTService as ChatGPTServiceType } from '../src/application/use-cases/ai/ChatGPTService';
-import { TestEnvService } from '../src/application/use-cases/env/TestEnvService';
+import type { ChatGPTService as ChatGPTServiceType } from '../src/infrastructure/external/ChatGPTService';
+import { TestEnvService } from '../src/infrastructure/config/TestEnvService';
 import type { PromptService } from '../src/application/interfaces/prompts/PromptService.interface';
 import type { LoggerFactory } from '../src/application/interfaces/logging/LoggerFactory.interface';
 
@@ -67,7 +67,7 @@ describe('ChatGPTService', () => {
       }),
     } as unknown as LoggerFactory;
     ({ ChatGPTService } = await import(
-      '../src/application/use-cases/ai/ChatGPTService'
+      '../src/infrastructure/external/ChatGPTService'
     ));
     service = new ChatGPTService(
       env,
