+++ test/Triggers.test.ts
@@ -2,7 +2,7 @@ import type { Context } from 'telegraf';
 import { describe, expect, it, vi } from 'vitest';
 
 import { DefaultDialogueManager } from '../src/application/use-cases/chat/DefaultDialogueManager';
-import { TestEnvService } from '../src/application/use-cases/env/TestEnvService';
+import { TestEnvService } from '../src/infrastructure/config/TestEnvService';
 import { MentionTrigger } from '../src/triggers/MentionTrigger';
 import { NameTrigger } from '../src/triggers/NameTrigger';
 import { ReplyTrigger } from '../src/triggers/ReplyTrigger';
