+++ vitest.config.ts
@@ -11,9 +11,9 @@ export default defineConfig({
         '**/index.ts',
         'src/migrate.ts',
         'src/container.ts',
-        'src/infrastructure/repositories/**',
-        'src/application/use-cases/env/**',
-        'src/bot/**',
+        'src/infrastructure/persistence/**',
+        'src/infrastructure/config/**',
+        'src/infrastructure/telegram/**',
         'src/triggers/**',
         'src/domain/entities/**',
         '**/*.interface.ts',
